+++ tuneuptechnology/client.py
@@ -1,19 +1,25 @@
 import requests
 
-VERSION = '2.1.0'
-API_BASE_URL = "https://app.tuneuptechnology.com/api/"
-
 
 class Client():
+    VERSION = '2.1.0'
+    API_BASE_URL = "https://app.tuneuptechnology.com/api/"
+
     @staticmethod
-    def response(data, endpoint):
+    def _make_http_request(data, endpoint):
         """Build the API request and return it to the method invoking it"""
         headers = {
-            'User-Agent': f'TuneupTechnologyApp/PythonClient/{VERSION}'
+            'User-Agent': f'TuneupTechnologyApp/PythonClient/{Client.VERSION}'
         }
-        request = requests.post(
-            endpoint,
-            data=data,
-            headers=headers,
-        )
+
+        try:
+            request = requests.post(
+                endpoint,
+                data=data,
+                headers=headers,
+                timeout=10,
+            )
+        except Exception as error:
+            raise Exception(error)
+
         return request